requirements.txt
selenium==4.10.0
webdriver-manager==3.8.6
requests==2.31.0
beautifulsoup4==4.12.2
pillow==10.0.0
python-dotenv==1.0.0
schedule==1.2.0

pytesseract==0.3.10
opencv-python-headless==4.8.0.74

lxml==4.9.3

concurrent-log-handler==0.9.24
fpdf==1.7.2

undetected-chromedriver==3.5.0

reportlab==4.0.4

numpy

boto3==1.28.0